Overhauled timing system by changing fixed timestep and time calculation
Current time of the song is now calculated from its time samples divided by the sample rate which makes it way more accurate instead of using .time property. Every system is now using that instead of the old .time. In addition player movement speed is now not tied to the fixedDeltaTime but the passed time of song, so if the player lags it should no longer cause the player to go offbeat since it follows the played time only.